requirements.txt
flask==2.0.2
tfhub-styletransfer-wrapper==1.1.3
flask-cors==3.0.10
gunicorn==20.1.0
Pillow==9.0.0
numpy==1.21.5
opencv-python-headless==4.5.5.62
